refactor: merge parse & element
