+++ src/Makefile
@@ -6,7 +6,7 @@
 #    By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/05/30 21:51:46 by minjungk          #+#    #+#              #
-#    Updated: 2023/06/07 07:07:16 by yonshin          ###   ########.fr        #
+#    Updated: 2023/06/09 22:55:36 by minjungk         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -65,7 +65,6 @@ MINIRT_SRCS = \
 	Element/Plane/plane.c \
 	Element/Sphere/sphere.c \
 	Element/Cylinder/cylinder.c \
-	parse/element.c \
 	parse/parse.c \
 	simulate/simulate.c \
 	simulate/hook_key_event.c \
